requirements.txt
bert-extractive-summarizer==0.7.1
blis==0.7.4
catalogue==2.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cymem==2.0.5
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
murmurhash==1.0.5
numpy==1.19.0
olefile==0.46
packaging==20.9
pathy==0.4.0
preshed==3.0.5
pydantic==1.7.3
pyparsing==2.4.7
regex==2020.11.13
requests==2.25.1
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.6.1
smart-open==3.0.0
spacy==3.0.5
spacy-legacy==3.0.1
srsly==2.4.0
thinc==8.0.2
threadpoolctl==2.1.0
tokenizers==0.10.1
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
tqdm==4.59.0
transformers==4.3.3
typer==0.3.2
urllib3==1.26.3
wasabi==0.8.2
Werkzeug==1.0.1
wincertstore==0.2
